+++ src/game/Object/UnitMeleeDamage.cpp
@@ -1,0 +1,551 @@
+/**
+ * MaNGOS is a full featured server for World of Warcraft, supporting
+ * the following clients: 1.12.x, 2.4.3, 3.3.5a, 4.3.4a and 5.4.8
+ *
+ * Copyright (C) 2005-2025 MaNGOS <https://www.getmangos.eu>
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ *
+ * World of Warcraft, and all World of Warcraft or Warcraft art, images,
+ * and lore are copyrighted by Blizzard Entertainment, Inc.
+ */
+
+/**
+ * @file UnitMeleeDamage.cpp
+ * @brief Cohesion split of Unit.cpp -- melee damage calculation and application (CalculateMeleeDamage / DealMeleeDamage).
+ *        Same `Unit` class; no behaviour change.
+ */
+
+#include "Unit.h"
+#include "Log.h"
+#include "Opcodes.h"
+#include "WorldPacket.h"
+#include "WorldSession.h"
+#include "World.h"
+#include "ObjectMgr.h"
+#include "ObjectGuid.h"
+#include "SpellMgr.h"
+#include "QuestDef.h"
+#include "Player.h"
+#include "Creature.h"
+#include "Spell.h"
+#include "Group.h"
+#include "SpellAuras.h"
+#include "MapManager.h"
+#include "ObjectAccessor.h"
+#include "CreatureAI.h"
+#include "TemporarySummon.h"
+#include "Formulas.h"
+#include "Pet.h"
+#include "Util.h"
+#include "Totem.h"
+#include "Vehicle.h"
+#include "BattleGround/BattleGround.h"
+#include "InstanceData.h"
+#include "OutdoorPvP/OutdoorPvP.h"
+#include "MapPersistentStateMgr.h"
+#include "GridNotifiersImpl.h"
+#include "CellImpl.h"
+#include "VMapFactory.h"
+#include "CreatureLinkingMgr.h"
+#include "GameTime.h"
+#ifdef ENABLE_ELUNA
+#include "LuaEngine.h"
+#include "ElunaConfig.h"
+#include "ElunaEventMgr.h"
+#endif /* ENABLE_ELUNA */
+
+// TODO for melee need create structure as in
+void Unit::CalculateMeleeDamage(Unit* pVictim, CalcDamageInfo* damageInfo, WeaponAttackType attackType /*= BASE_ATTACK*/)
+{
+    damageInfo->attacker         = this;
+    damageInfo->target           = pVictim;
+    damageInfo->damageSchoolMask = GetMeleeDamageSchoolMask();
+    damageInfo->attackType       = attackType;
+    damageInfo->damage           = 0;
+    damageInfo->cleanDamage      = 0;
+    damageInfo->absorb           = 0;
+    damageInfo->resist           = 0;
+    damageInfo->blocked_amount   = 0;
+
+    damageInfo->TargetState      = VICTIMSTATE_UNAFFECTED;
+    damageInfo->HitInfo          = HITINFO_NORMALSWING;
+    damageInfo->procAttacker     = PROC_FLAG_NONE;
+    damageInfo->procVictim       = PROC_FLAG_NONE;
+    damageInfo->procEx           = PROC_EX_NONE;
+    damageInfo->hitOutCome       = MELEE_HIT_EVADE;
+
+    if (!this || !pVictim)
+    {
+        return;
+    }
+    if (!this->IsAlive() || !pVictim->IsAlive())
+    {
+        return;
+    }
+
+    // Select HitInfo/procAttacker/procVictim flag based on attack type
+    switch (attackType)
+    {
+        case BASE_ATTACK:
+            damageInfo->procAttacker = PROC_FLAG_SUCCESSFUL_MELEE_HIT;
+            damageInfo->procVictim   = PROC_FLAG_TAKEN_MELEE_HIT;
+            damageInfo->HitInfo      = HITINFO_NORMALSWING2;
+            break;
+        case OFF_ATTACK:
+            damageInfo->procAttacker = PROC_FLAG_SUCCESSFUL_MELEE_HIT | PROC_FLAG_SUCCESSFUL_OFFHAND_HIT;
+            damageInfo->procVictim   = PROC_FLAG_TAKEN_MELEE_HIT;//|PROC_FLAG_TAKEN_OFFHAND_HIT // not used
+            damageInfo->HitInfo = HITINFO_LEFTSWING;
+            break;
+        case RANGED_ATTACK:
+            damageInfo->procAttacker = PROC_FLAG_SUCCESSFUL_RANGED_HIT;
+            damageInfo->procVictim   = PROC_FLAG_TAKEN_RANGED_HIT;
+            damageInfo->HitInfo = HITINFO_UNK3;             // test (dev note: test what? HitInfo flag possibly not confirmed.)
+            break;
+        default:
+            break;
+    }
+
+    // Physical Immune check
+    if (damageInfo->target->IsImmuneToDamage(damageInfo->damageSchoolMask))
+    {
+        damageInfo->HitInfo       |= HITINFO_NORMALSWING;
+        damageInfo->TargetState    = VICTIMSTATE_IS_IMMUNE;
+
+        damageInfo->procEx |= PROC_EX_IMMUNE;
+        damageInfo->damage         = 0;
+        damageInfo->cleanDamage    = 0;
+        return;
+    }
+    uint32 damage = CalculateDamage(damageInfo->attackType, false);
+    // Add melee damage bonus
+    damage = MeleeDamageBonusDone(damageInfo->target, damage, damageInfo->attackType);
+    damage = damageInfo->target->MeleeDamageBonusTaken(this, damage, damageInfo->attackType);
+
+    // Calculate armor reduction
+    uint32 armor_affected_damage = CalcNotIgnoreDamageReduction(damage, damageInfo->damageSchoolMask);
+    damageInfo->damage = damage - armor_affected_damage + CalcArmorReducedDamage(damageInfo->target, armor_affected_damage);
+    damageInfo->cleanDamage += damage - damageInfo->damage;
+
+    damageInfo->hitOutCome = RollMeleeOutcomeAgainst(damageInfo->target, damageInfo->attackType);
+
+    // Disable parry or dodge for ranged attack
+    if (damageInfo->attackType == RANGED_ATTACK)
+    {
+        if (damageInfo->hitOutCome == MELEE_HIT_PARRY)
+        {
+            damageInfo->hitOutCome = MELEE_HIT_NORMAL;
+        }
+        if (damageInfo->hitOutCome == MELEE_HIT_DODGE)
+        {
+            damageInfo->hitOutCome = MELEE_HIT_MISS;
+        }
+    }
+
+    switch (damageInfo->hitOutCome)
+    {
+        case MELEE_HIT_EVADE:
+        {
+            damageInfo->HitInfo    |= HITINFO_MISS | HITINFO_SWINGNOHITSOUND;
+            damageInfo->TargetState = VICTIMSTATE_EVADES;
+
+            damageInfo->procEx |= PROC_EX_EVADE;
+            damageInfo->damage = 0;
+            damageInfo->cleanDamage = 0;
+            return;
+        }
+        case MELEE_HIT_MISS:
+        {
+            damageInfo->HitInfo    |= HITINFO_MISS;
+            damageInfo->TargetState = VICTIMSTATE_UNAFFECTED;
+
+            damageInfo->procEx |= PROC_EX_MISS;
+            damageInfo->damage = 0;
+            damageInfo->cleanDamage = 0;
+            break;
+        }
+        case MELEE_HIT_NORMAL:
+            damageInfo->TargetState = VICTIMSTATE_NORMAL;
+            damageInfo->procEx |= PROC_EX_NORMAL_HIT;
+            break;
+        case MELEE_HIT_CRIT:
+        {
+            damageInfo->HitInfo     |= HITINFO_CRITICALHIT;
+            damageInfo->TargetState  = VICTIMSTATE_NORMAL;
+
+            damageInfo->procEx |= PROC_EX_CRITICAL_HIT;
+            // Crit bonus calc
+            damageInfo->damage += damageInfo->damage;
+            int32 mod = 0;
+            // Apply SPELL_AURA_MOD_ATTACKER_RANGED_CRIT_DAMAGE or SPELL_AURA_MOD_ATTACKER_MELEE_CRIT_DAMAGE
+            if (damageInfo->attackType == RANGED_ATTACK)
+            {
+                mod += damageInfo->target->GetTotalAuraModifier(SPELL_AURA_MOD_ATTACKER_RANGED_CRIT_DAMAGE);
+            }
+            else
+            {
+                mod += damageInfo->target->GetTotalAuraModifier(SPELL_AURA_MOD_ATTACKER_MELEE_CRIT_DAMAGE);
+            }
+
+            mod += GetTotalAuraModifierByMiscMask(SPELL_AURA_MOD_CRIT_DAMAGE_BONUS, SPELL_SCHOOL_MASK_NORMAL);
+
+            uint32 crTypeMask = damageInfo->target->GetCreatureTypeMask();
+
+            // Increase crit damage from SPELL_AURA_MOD_CRIT_PERCENT_VERSUS
+            mod += GetTotalAuraModifierByMiscMask(SPELL_AURA_MOD_CRIT_PERCENT_VERSUS, crTypeMask);
+            if (mod != 0)
+            {
+                damageInfo->damage = int32((damageInfo->damage) * float((100.0f + mod) / 100.0f));
+            }
+
+            // Resilience - reduce crit damage
+            uint32 reduction_affected_damage = CalcNotIgnoreDamageReduction(damageInfo->damage, damageInfo->damageSchoolMask);
+            uint32 resilienceReduction;
+            if (attackType != RANGED_ATTACK)
+            {
+                resilienceReduction = pVictim->GetMeleeCritDamageReduction(reduction_affected_damage);
+            }
+            else
+            {
+                resilienceReduction = pVictim->GetRangedCritDamageReduction(reduction_affected_damage);
+            }
+
+            damageInfo->damage      -= resilienceReduction;
+            damageInfo->cleanDamage += resilienceReduction;
+            break;
+        }
+        case MELEE_HIT_PARRY:
+            damageInfo->TargetState  = VICTIMSTATE_PARRY;
+            damageInfo->procEx      |= PROC_EX_PARRY;
+            damageInfo->cleanDamage += damageInfo->damage;
+            damageInfo->damage = 0;
+            break;
+
+        case MELEE_HIT_DODGE:
+            damageInfo->TargetState  = VICTIMSTATE_DODGE;
+            damageInfo->procEx      |= PROC_EX_DODGE;
+            damageInfo->cleanDamage += damageInfo->damage;
+            damageInfo->damage = 0;
+            break;
+        case MELEE_HIT_BLOCK:
+        {
+            damageInfo->TargetState = VICTIMSTATE_NORMAL;
+            damageInfo->HitInfo    |= HITINFO_BLOCK;
+            damageInfo->procEx     |= PROC_EX_BLOCK;
+            damageInfo->blocked_amount = damageInfo->target->GetShieldBlockValue();
+
+            // Target has a chance to double the blocked amount if it has SPELL_AURA_MOD_BLOCK_CRIT_CHANCE
+            if (roll_chance_i(pVictim->GetTotalAuraModifier(SPELL_AURA_MOD_BLOCK_CRIT_CHANCE)))
+            {
+                damageInfo->blocked_amount *= 2;
+            }
+
+            if (damageInfo->blocked_amount >= damageInfo->damage)
+            {
+                damageInfo->TargetState = VICTIMSTATE_BLOCKS;
+                damageInfo->blocked_amount = damageInfo->damage;
+                damageInfo->procEx |= PROC_EX_FULL_BLOCK;
+            }
+            else
+            {
+                damageInfo->procEx |= PROC_EX_NORMAL_HIT;    // Partial blocks can still cause attacker procs
+            }
+
+            damageInfo->damage      -= damageInfo->blocked_amount;
+            damageInfo->cleanDamage += damageInfo->blocked_amount;
+            break;
+        }
+        case MELEE_HIT_GLANCING:
+        {
+            damageInfo->HitInfo |= HITINFO_GLANCING;
+            damageInfo->TargetState = VICTIMSTATE_NORMAL;
+            damageInfo->procEx |= PROC_EX_NORMAL_HIT;
+            float reducePercent = 1.0f;                     // damage factor
+            // calculate base values and mods
+            float baseLowEnd = 1.3f;
+            float baseHighEnd = 1.2f;
+            switch (getClass())                             // lowering base values for casters
+            {
+                case CLASS_SHAMAN:
+                case CLASS_PRIEST:
+                case CLASS_MAGE:
+                case CLASS_WARLOCK:
+                case CLASS_DRUID:
+                    baseLowEnd  -= 0.7f;
+                    baseHighEnd -= 0.3f;
+                    break;
+            }
+
+            float maxLowEnd = 0.6f;
+            switch (getClass())                             // upper for melee classes
+            {
+                case CLASS_WARRIOR:
+                case CLASS_ROGUE:
+                    maxLowEnd = 0.91f;                      // If the attacker is a melee class then instead the lower value of 0.91
+            }
+
+            // calculate values
+            int32 diff = damageInfo->target->GetDefenseSkillValue() - GetWeaponSkillValue(damageInfo->attackType);
+            float lowEnd  = baseLowEnd - (0.05f * diff);
+            float highEnd = baseHighEnd - (0.03f * diff);
+
+            // apply max/min bounds
+            if (lowEnd < 0.01f)                             // the low end must not go bellow 0.01f
+            {
+                lowEnd = 0.01f;
+            }
+            else if (lowEnd > maxLowEnd)                    // the smaller value of this and 0.6 is kept as the low end
+            {
+                lowEnd = maxLowEnd;
+            }
+
+            if (highEnd < 0.2f)                             // high end limits
+            {
+                highEnd = 0.2f;
+            }
+            if (highEnd > 0.99f)
+            {
+                highEnd = 0.99f;
+            }
+
+            if (lowEnd > highEnd)                           // prevent negative range size
+            {
+                lowEnd = highEnd;
+            }
+
+            reducePercent = lowEnd + rand_norm_f() * (highEnd - lowEnd);
+
+            damageInfo->cleanDamage += damageInfo->damage - uint32(reducePercent *  damageInfo->damage);
+            damageInfo->damage   = uint32(reducePercent *  damageInfo->damage);
+            break;
+        }
+        case MELEE_HIT_CRUSHING:
+        {
+            damageInfo->HitInfo     |= HITINFO_CRUSHING;
+            damageInfo->TargetState  = VICTIMSTATE_NORMAL;
+            damageInfo->procEx |= PROC_EX_NORMAL_HIT;
+            // 150% normal damage
+            damageInfo->damage += (damageInfo->damage / 2);
+            break;
+        }
+        default:
+
+            break;
+    }
+
+    // only from players
+    if (GetTypeId() == TYPEID_PLAYER)
+    {
+        uint32 reduction_affected_damage = CalcNotIgnoreDamageReduction(damageInfo->damage, damageInfo->damageSchoolMask);
+        uint32 resilienceReduction;
+        if (attackType != RANGED_ATTACK)
+        {
+            resilienceReduction = pVictim->GetMeleeDamageReduction(reduction_affected_damage);
+        }
+        else
+        {
+            resilienceReduction = pVictim->GetRangedDamageReduction(reduction_affected_damage);
+        }
+        damageInfo->damage      -= resilienceReduction;
+        damageInfo->cleanDamage += resilienceReduction;
+    }
+
+    // Calculate absorb resist
+    if (int32(damageInfo->damage) > 0)
+    {
+        damageInfo->procVictim |= PROC_FLAG_TAKEN_ANY_DAMAGE;
+
+        // Calculate absorb & resists
+        uint32 absorb_affected_damage = CalcNotIgnoreAbsorbDamage(damageInfo->damage, damageInfo->damageSchoolMask);
+        damageInfo->target->CalculateDamageAbsorbAndResist(this, damageInfo->damageSchoolMask, DIRECT_DAMAGE, absorb_affected_damage, &damageInfo->absorb, &damageInfo->resist, true);
+        damageInfo->damage -= damageInfo->absorb + damageInfo->resist;
+        if (damageInfo->absorb)
+        {
+            damageInfo->HitInfo |= HITINFO_ABSORB;
+            damageInfo->procEx |= PROC_EX_ABSORB;
+        }
+        if (damageInfo->resist)
+        {
+            damageInfo->HitInfo |= HITINFO_RESIST;
+        }
+    }
+    else // Umpossible get negative result but....
+    {
+        damageInfo->damage = 0;
+    }
+}
+
+/**
+ * @brief Applies prepared melee damage and related proc logic.
+ *
+ * @param damageInfo The prepared melee damage information.
+ * @param durabilityLoss True to apply durability loss rules.
+ */
+void Unit::DealMeleeDamage(CalcDamageInfo* damageInfo, bool durabilityLoss)
+{
+    if (damageInfo == 0)
+    {
+        return;
+    }
+    Unit* pVictim = damageInfo->target;
+
+    if (!this || !pVictim)
+    {
+        return;
+    }
+
+    if (!pVictim->IsAlive() || pVictim->IsTaxiFlying() || (pVictim->GetTypeId() == TYPEID_UNIT && ((Creature*)pVictim)->IsInEvadeMode()))
+    {
+        return;
+    }
+
+    // You don't lose health from damage taken from another player while in a sanctuary
+    // You still see it in the combat log though
+    if (!IsAllowedDamageInArea(pVictim))
+    {
+        return;
+    }
+
+    // Hmmmm dont like this emotes client must by self do all animations
+    if (damageInfo->HitInfo & HITINFO_CRITICALHIT)
+    {
+        pVictim->HandleEmoteCommand(EMOTE_ONESHOT_WOUNDCRITICAL);
+    }
+    if (damageInfo->blocked_amount && damageInfo->TargetState != VICTIMSTATE_BLOCKS)
+    {
+        pVictim->HandleEmoteCommand(EMOTE_ONESHOT_PARRYSHIELD);
+    }
+
+    // This seems to reduce the victims time until next attack if your attack was parried
+    if (damageInfo->TargetState == VICTIMSTATE_PARRY)
+    {
+        // Get attack timers
+        float offtime  = float(pVictim->getAttackTimer(OFF_ATTACK));
+        float basetime = float(pVictim->getAttackTimer(BASE_ATTACK));
+        // Reduce attack time
+        if (pVictim->haveOffhandWeapon() && offtime < basetime)
+        {
+            float percent20 = pVictim->GetAttackTime(OFF_ATTACK) * 0.20f;
+            float percent60 = 3.0f * percent20;
+            if (offtime > percent20 && offtime <= percent60)
+            {
+                pVictim->setAttackTimer(OFF_ATTACK, uint32(percent20));
+            }
+            else if (offtime > percent60)
+            {
+                offtime -= 2.0f * percent20;
+                pVictim->setAttackTimer(OFF_ATTACK, uint32(offtime));
+            }
+        }
+        else
+        {
+            float percent20 = pVictim->GetAttackTime(BASE_ATTACK) * 0.20f;
+            float percent60 = 3.0f * percent20;
+            if (basetime > percent20 && basetime <= percent60)
+            {
+                pVictim->setAttackTimer(BASE_ATTACK, uint32(percent20));
+            }
+            else if (basetime > percent60)
+            {
+                basetime -= 2.0f * percent20;
+                pVictim->setAttackTimer(BASE_ATTACK, uint32(basetime));
+            }
+        }
+    }
+
+    // Call default DealDamage
+    CleanDamage cleanDamage(damageInfo->cleanDamage, damageInfo->attackType, damageInfo->hitOutCome);
+    DealDamage(pVictim, damageInfo->damage, &cleanDamage, DIRECT_DAMAGE, damageInfo->damageSchoolMask, NULL, durabilityLoss);
+
+    // If this is a creature and it attacks from behind it has a probability to daze it's victim
+    if ((damageInfo->hitOutCome == MELEE_HIT_CRIT || damageInfo->hitOutCome == MELEE_HIT_CRUSHING || damageInfo->hitOutCome == MELEE_HIT_NORMAL || damageInfo->hitOutCome == MELEE_HIT_GLANCING) &&
+        GetTypeId() != TYPEID_PLAYER && !((Creature*)this)->GetCharmerOrOwnerGuid() && !pVictim->HasInArc(M_PI_F, this))
+    {
+        // -probability is between 0% and 40%
+        // 20% base chance
+        float Probability = 20.0f;
+
+        // there is a newbie protection, at level 10 just 7% base chance; assuming linear function
+        if (pVictim->getLevel() < 30)
+        {
+            Probability = 0.65f * pVictim->getLevel() + 0.5f;
+        }
+
+        uint32 VictimDefense = pVictim->GetDefenseSkillValue();
+        uint32 AttackerMeleeSkill = GetUnitMeleeSkill();
+
+        Probability *= AttackerMeleeSkill / (float)VictimDefense;
+
+        if (Probability > 40.0f)
+        {
+            Probability = 40.0f;
+        }
+
+        if (roll_chance_f(Probability))
+        {
+            CastSpell(pVictim, 1604, true);
+        }
+    }
+
+    // If not miss
+    if (!(damageInfo->HitInfo & HITINFO_MISS))
+    {
+        // on weapon hit casts
+        if (GetTypeId() == TYPEID_PLAYER && pVictim->IsAlive())
+        {
+            ((Player*)this)->CastItemCombatSpell(pVictim, damageInfo->attackType);
+        }
+
+        // victim's damage shield
+        std::set<Aura*> alreadyDone;
+        AuraList const& vDamageShields = pVictim->GetAurasByType(SPELL_AURA_DAMAGE_SHIELD);
+        for (AuraList::const_iterator i = vDamageShields.begin(); i != vDamageShields.end();)
+        {
+            if (alreadyDone.find(*i) == alreadyDone.end())
+            {
+                alreadyDone.insert(*i);
+                uint32 damage = (*i)->GetModifier()->m_amount;
+                SpellEntry const* i_spellProto = (*i)->GetSpellProto();
+                // Calculate absorb resist ??? no data in opcode for this possibly unable to absorb or resist?
+                // uint32 absorb;
+                // uint32 resist;
+                // CalcAbsorbResist(pVictim, SpellSchools(spellProto->School), SPELL_DIRECT_DAMAGE, damage, &absorb, &resist);
+                // damage-=absorb + resist;
+
+                pVictim->DealDamageMods(this, damage, NULL);
+
+                uint32 targetHealth = GetHealth();
+                uint32 overkill = damage > targetHealth ? damage - targetHealth : 0;
+
+                WorldPacket data(SMSG_SPELLDAMAGESHIELD, (8 + 8 + 4 + 4 + 4 + 4));
+                data << pVictim->GetObjectGuid();
+                data << GetObjectGuid();
+                data << uint32(i_spellProto->Id);
+                data << uint32(damage);                  // Damage
+                data << uint32(overkill);                // Overkill
+                data << uint32(i_spellProto->SchoolMask);
+                pVictim->SendMessageToSet(&data, true);
+
+                pVictim->DealDamage(this, damage, 0, SPELL_DIRECT_DAMAGE, GetSpellSchoolMask(i_spellProto), i_spellProto, true);
+
+                i = vDamageShields.begin();
+            }
+            else
+            {
+                ++i;
+            }
+        }
+    }
+}
